requirement.txt
asgiref==3.2.5
astroid==2.3.3
colorama==0.4.3
Django==2.2
django-bootstrap-modal-forms==1.5.0
django-crispy-forms==1.9.0
django-multiselectfield==0.1.12
django-phone-field==1.8.0
django-widget-tweaks==1.4.8
gunicorn==20.0.4
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
Pillow==7.0.0
psycopg2==2.8.4
pylint==2.4.4
pylint-django==2.0.14
pylint-plugin-utils==0.6
pytz==2019.3
six==1.14.0
sqlparse==0.3.1
typed-ast==1.4.1
whitenoise==5.0.1
wrapt==1.11.2
